requirements.txt
fastapi
uvicorn[standard]
openai
pinecone-client
python-multipart
streamlit
PyPDF2
python-docx
pydantic
requests
python-dotenv
